+++ CHANGELOG.md
@@ -25,12 +25,16 @@ Changes are kept under subsections:
 Only major refactoring improvements or functional changes should be noted.
 
 --------------------------------------------------------------------------------
-## [1.4.2-bulk] - 2025-02-xx - Refactor Events Requests
-#### Fixed
+## [1.5.2] - 2025-02-13 - Stats reports
+#### New
+- Plot NConf distribution
+- Plot Events distribution
+- Cached data mechanism
+
+# [1.4.2] - 2025-02-10 - Refactor Events Requests
 #### Changed
 - NConf calculation optimized
 - Missions sorting on catalogs page
-#### New
 
 ## [1.3.3] - 2025-02-07 - Partial bulk by CLI
 #### Fixed